+++ samples/Calculator/Contract/FileCalculationRequest.cs
@@ -1,9 +1,10 @@
 ﻿using System.IO;
+using XRoadLib;
 using XRoadLib.Serialization;
 
 namespace Calculator.Contract
 {
-    public class FileCalculationRequest : XRoadSerializable
+    public class FileCalculationRequest : XRoadSerializable, IXRoadRequest<Stream>
     {
         public Stream InputFile { get; set; }
     }