+++ src/XRoadLib.Extensions.Http/Services/IXRoadService.cs
@@ -8,10 +8,10 @@ namespace XRoadLib.Extensions.Http.Services
         /// <summary>
         /// Executes X-Road operation on endpoint specified by WebRequest parameter.
         /// </summary>
-        /// <param name="body">Soap body part of outgoing serialized X-Road message.</param>
+        /// <param name="request">Soap body part of outgoing serialized X-Road message.</param>
         /// <param name="header">Soap header part of outgoing serialized X-Road message.</param>
         /// <param name="options">Additional options to configure service call execution.</param>
         /// <returns>Deserialized value of X-Road response message Soap body.</returns>
-        Task<XRoadResponse> ExecuteAsync(object body, ISoapHeader header, ServiceExecutionOptions options = null);
+        Task<XRoadResponse<TResult>> ExecuteAsync<TResult>(IXRoadRequest<TResult> request, ISoapHeader header, ServiceExecutionOptions options = null);
     }
 }