+++ src/XRoadLib/IXRoadRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace XRoadLib
+{
+    public interface IXRoadRequest<TResponse>
+    {
+        
+    }
+}