+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 * Removed generic ServiceManager (HeaderDefinition provides header implementation from this point on).
 * Replace HttpWebRequest based client request with HttpClient based implementation.
+* Add IXRoadRequest<TResponse>, IXRoadRequestHandler<TRequest, TResponse> type for improved typing support.
 
 #### 1.4.1 - 13.12.2020